Fix to circular errors
